[hack] Exit if lockfile gets deleted
Summary: Previously, we would attempt to recreate the lockfile, but that
doesn't work very well if the socket file got deleted too -- the client
would then not be able to connect, but it wouldn't be able to start up a
new server since the lockfile would still be held.

Also update the mtime of the lockfile periodically so this failure mode
is less likely to happen.

We should probably also check if the socket file gets deleted and exit in that
case too, but that'd require a bit more refactoring.

Reviewed By: @jwatzman

Differential Revision: D2283846