use a variant to represent string/number literal values
Summary: it was pretty hacky to represent refined `StrT` literal values as `Some "truthy"`, since that is not distinguishable from literal `"truthy"`. this leads to issues like

```
function foo(obj, x: ?string) {
  if (x) {
    // now we think x === "truthy"
    obj[x] = ...; // adds a "truthy" prop to obj
  }
}
```

this diff uses a variant to tell the difference.

Technically, we don't need `Falsy` because strings and numbers only have one falsy value, so we could do `Literal ""`, but it didn't really make anything more gross and I think is a little less clever to have `Falsy`.

Reviewed By: @gabelevi

Differential Revision: D2287603