Remove sysConfig
Summary: this diff eliminates flow's override of sysConfig.ml, which is super hacky. if ocamlbuild builds things in the wrong order, or the dependencies change subtly, it'll see the wrong version of sysConfig (or both, and error).

Instead, moves filename generation out of `Tmp` or `Lock` and into `GlobalConfig` (for hack) and `FlowConfig` (for flow), so that the modules in `utils` are agnostic. I also like that it gives a central place to manage all of the files that get created.

Reviewed By: @bhosmer

Differential Revision: D2284919