Declare overloaded versions of fs.readFile and fs.readFileSync
Summary: If no options are provided, data is a Buffer.
If options are provided without an encoding set, data is a Buffer.
If options are provided and an encoding is specified, data is a string.

If the second parameter is a string, data is a Buffer. This is the "old"
API, so it's undocumented, but it is still supported and does the right
thing.
Closes https://github.com/facebook/flow/pull/653

Reviewed By: @mroch

Differential Revision: D2270834

Pulled By: @jeffmo