Basic support for unsound getters & setters
Summary: It's going to be really hard to support getters and setters in any
sort of sound way. Our refinement system relies on knowing when a function is
being called in order to havok the non-local variable refinements, but due to
the way we split work among the workers, we don't yet know which properties
have getters & setters.

This support is by default disabled. You can turn it on with a config option
`unsafe.enable_getters_and_setters`.

I also began to notice that this feature is actually more complicated than I
anticipated, so I haven't really covered all the weird corner cases terribly
well. Proper support would probably involve keeping a separate type for
reading/writing properties. There might also be compromises, like making all
setter-only properties optional, in case someone reads them (which returns
undefined).

This work also made me realize that duplicate properties are allowed in ES6
classes and objects. The parser currently disallows them. I don't know if we
really want to allow this stuff in Flow, but if not we should move that out of
the parser and into the typechecker.

Reviewed By: @jeffmo

Differential Revision: D2258879