Limit json printing by depth
Summary: This diff allows us to limit the depth of the JSON we generate in
`constraint_js.ml`. Basically I threaded a `json_cx` through everywhere and
wrapped each json-generating function with something that would check and
decrement the depth. When we reach the depth limit, we we stop recursing and
return null.

The one thing that doesn't respect depth is the json_of_reason call. I
considered adding support, but a partially printed reason seemed kind of silly.

Reviewed By: @mroch

Differential Revision: D2292605