make temp dir configurable
Summary: It would be useful during development to run multiple versions of flow (`flow` and `flow-dev`, for instance), even though we generally want to only run a single version. This restriction is currently enforced by `FlowConfig.lock_file` (and friends) which hardcodes the temp dir path so that the pid/lock/socket files that the client and server use to communicate exist in exactly one place, preventing two instances from running on the same root dir.

By changing the base temp dir, we can create an entire separate environment for these files, so that `flow` and `flow-dev` can't see each other. This diff adds a `--temp-dir` command line option and threads that around to all of the places that create lock files and sockets.

Reviewed By: @gabelevi

Differential Revision: D2277560