[hack] Treat enum members as abstract types
Summary: Previously, we were treating enums as special class instances,
which led to some weird and incorrect semantics. In particular, the
following would typecheck:

  enum Foo {
    X = 1;
  }

  function f() {
    $q = Foo::X;
    $q::getValues();
  }

However, since `$q` is an int and not an object, this method call would
fail at runtime.

This also allows us to have enums of classname work as expected (see
unit test.)

Reviewed By: @dlreeves

Differential Revision: D2294119