obj lit assign
Summary: Fix for issue described in T7810506: there it's described in the context
of definite assignment of object properties, but it's actually more
general. Basically, when an object literal is used to initialize an
lvalue (say a variable or object property) in such a way that the
literal's type then becomes the lvalue's observable type to downstream
code, the optional properties would become inaccessible, causing errors
when subsequent assignments tried to add the optional properties to the
lvalue. (See tests/optional_props/test3.js for examples.)

We had gone to some trouble to account for this case in other contexts
by noting the flow of a sealed object type to an optional-property-carrying
object type and adding shadow (deferred-lookup) properties. But this doesn't
handle this subsequent-write-to-a-sealed-type problem, and arguably isn't
quite right for other cases either, since it omits the possibility of the
property being undefined. Now we add full-fledged properties of the given
optional type, which fixes the present bug and improves checking in cases
which previously neglected to error on the optional property's
potential to be undefined.

Reviewed By: @gabelevi

Differential Revision: D2280947