[hack] Add a notion of aliasing to auto-import HH\ things
Summary: Add a notion of an name alias map that we use for function and class names, mostly to pull certain things from the HH\ namespace into the root namespace. The map is copied from the ad-hoc places in the HHVM parser that it currently exists. The new `fmt_name` function that applies the aliases is used basically everwhere `strip_ns` used to be used for emitting things.

Reviewed By: @int3

Differential Revision: D2310618