verify dictionary properties as lower bound
Summary: if you have a dictionary `x` and access `x.foo`, we currently set the dictionary's type as an upper bound on `x.foo`, but no lower bound. if `x.foo` is never explicitly set (because it's a dictionary, we may never know), then no lower bound is ever set, and the users of `x.foo` are effectively `any`.

this diff adds dictionary handling to `ensure_prop_for_read`. if the prop doesn't exist as a standalone prop, then the dictionary value type is returned instead.

also fixes several other dictionary-related bugs:
- Object.prototype methods are allowed on records
- made dictionaries a subtype of structs if all of the struct's props match the indexer, carrying on the assumption that we make everywhere else that a dictionary has every prop
- relatedly, made dictionaries work with structural subtyping (`propTest5` in the tests failed before)

Reviewed By: @gabelevi

Differential Revision: D2310258