Create binding for named class expressions
If a class expression specifies a name, that name serves as a binding to
the class itself within the expression only.

I initially attempted to do this by converting the `Ast.Class` into a
`Ast.Statement.ClassDeclaration`, then running that through
`statement_decl` and `toplevels`, but in order to return the type from
the `expression` function, I needed to read the entry from the
environment, which doesn't work if the class expr is anonymous.

So, instead, I copied in the relevant parts of `statement_decl` and
`toplevels` into this part of the code. This kind of manual env
management inside of `expression` isn't unprecedented, but I'm open to
suggestions to improve this.

Fixes #699