[hack] Build system for Hack on Windows
Summary: This PR contains a build-system for Hack on windows. It relies on OCPWin (as described in `hphp/hack/README.win32`). This alternative build system may also be used on Linux and OS X (see `hphp/hack/README.ocp-build`).

*Warnings*: on Windows, the resulting binaries are not working yet, all tests are known to fail.

The main file describing the build process is `hphp/hack/src/build.ocp`. It defines a set of libraries and programs. The library architecture has been sketched with our young knownledge about Hack. If someone with more experience could have a look at it, it would be appreciated.

Some additional descriptions specific to Hack (i.e. not required when build Flow) may be found in the following files:

- `hphp/hack/build.ocp`
- `hphp/hack/src/h2tp/build.ocp`

While in the long term it might require some efforts to maintain two build systems, for now we take care not to modify (and not to conflict with) the current build system.
Closes https://github.com/facebook/hhvm/pull/5775

Reviewed By: @jwatzman

Differential Revision: D2268695

Pulled By: @gabelevi