[hack] Avoid zipping lists unnecessarily
Summary: In general, I'm trying to remove calls to functions that can
potentially raise exceptions, because they are a source of potential
surprise crashes.

Zipping throws an exception if the lists are not of the same length.
Most of the time they actually are of the same length -- we just need to
convince the typechecker of this fact.

Reviewed By: @dlreeves

Differential Revision: D2327293