[hack] Fix cmake build breakage caused by using Bytes module
Summary: Apparently the version of ocaml we use for our build machines for the cmake build doesn't have the Bytes module. I landed a use of Bytes in a patch earlier to fix a different build break resulting from a lack of the Str regexp module in js_of_ocaml. Fix this by going back to the regexp version with different implementations for ocaml and js land.

Reviewed By: @int3, @JoelMarcey

Differential Revision: D2340274