[hack] Fix autocomplete crashing hh_server
Summary: Because our error system is structured around a global module, all code that adds errors must be inside Errors.do_ or Errors.ignore_ call - otherwise nothing will clean them and they will accumulate forever, leaking memory (probably not important in server mode, because the errors are added by short lived worker processes, but it was a problem in js IDE code).

To prevent it D1571225 added an assertion that we've discovered is firing thanks to logging from D2313971.

The reason: it might have been true that none of the code in autocompleteService.ml could add errors, but it's not the case anymore - {D1982724} added calls to Typing_utils.localize that can do this. There are possibly others, but this is the one I've found. I fix it by wrapping the main entry point to this whole thing into Errors.ignore_.

Leaving in logging added in D2336478 that was supposed to help as debug this, because I have a feeling that this story will repeat itself.

Reviewed By: @int3

Differential Revision: D2345725