[hack] Implements `hh_server --daemon` on Windows.
Summary: In the absence of 'fork' on Windows, we have three distinct entry points. The default one is the 'deamon_main' function. Two alternate entry points may be selected by setting the environment variable "HH_SERVER_MAIN": they are named "Monitor" and "Daemon". (Work-in-progress: later we will add a third entry point for slaves)

They correspond to the two successive 'forks' on Linux :
- Monitor: open the log file and create the Deamon.
- Deamon: simply execute 'hh_server' (ignoring the '-d' option).

On Unix, we do not use this alternate entry points but we use 'fork' as usual.
Closes https://github.com/facebook/hhvm/pull/5930

Reviewed By: @int3

Differential Revision: D2342443

Pulled By: @jwatzman