not_found
Summary: Fix for Not_found arising from use of in-scope type params.
Basically the old setup wasn't equipped to handle a difference
between an instance's type_args and arg_polarity maps. With the
new handling of in-scope type params, type_args may hold args to
params from outer scopes. Long-term we need access to the polarities
of those params as well, but since polarity isn't exposed in language
yet, and the only non-neutral params aren't in outer scopes, we should
be okay defaulting to neutral.

Reviewed By: @avikchaudhuri

Differential Revision: D2348468