[hack] Fix race condition in tailing log reporter
Summary: Fix potential race condition in Hack server tailing logger. It's possible that the client and server race to the log file and so the client reports on an old log file before the server has a chance to write new messages. Since the messages we look at all have date stamps, the fix is to only report messages newer than when the client starts.

Also, replace '[]' when there are no log messages with '[no info yet]'

Reviewed By: @dabek

Differential Revision: D2337932