deprecate old error message format only available when passing --from
Summary: `flow status` and `flow check-contents` use a different, legacy error format when the `--from` flag was passed but `--json` was not.

this format is deprecated and rarely tested (in fact, it's currently broken), so refactored to make that obvious and clean up the non-deprecated paths. also, restricted this format to vim and emacs, so that adding `--from` to other, non-legacy client calls won't spread this bad format.

Reviewed By: @jeffmo

Differential Revision: D2334721