fix position of Object.freeze
Summary: `Object.freeze` both mutates and returns the object passed to it. When the argument is an object literal, we handle it specially so that we can mark the `ObjT` as frozen.

Initially I made the result use the location of the argument since it is mutated... or perhaps just by accident. Now I think it makes more sense for it to work like every other function, and return a type that points at the entire call.

For example, consider this normal function:

```
function foo(x: string): string {
  return x;
}

var y = foo("bar");
//      ^^^^^^^^^^   <- errors here, not on lines 1 or 2
(y: number);
```

Reviewed By: @jeffmo

Differential Revision: D2359393