Add token-streaming support to parser + `flow ast --tokens`
Summary: This adds support to the parser for specifying a token sink (callback that receives tokens as parsing progresses).

Additionally this exposes the token list to the command line via `flow ast --tokens`, which just tacks a list of "token" objects on the toplevel object that is printed. Having access to the raw token list is useful for tools that are interested in the concrete syntax more so than the AST (like syntax highlighters, code formatters, etc).

Reviewed By: @gabelevi

Differential Revision: D2329711