[hack] Rework the representation of strings
Summary: Rework the representation of strings such that literal strings (whether from single or double quotes) are stored as String and are unescaped. String2 then loses its string argument and contains only the list of expressions used to construct it (which is extended to include the literal string parts). That is, String2 is now essentially the "concat these exprs" operator.

Reviewed By: @int3

Differential Revision: D2349624