[hack] Revert logging code to print plain lines instead of fancy ellipses
Summary: The ellipses were initially intended to act as progress indicators in
some sort of hh_client `tail`, but @​bpr's implementation of the tailer didn't
end up making use of it. Remove so that code is cleaner and so that we can add
more log statements inside the individual phases without it messing up the
output.

Reviewed By: @dabek

Differential Revision: D2365012