Allow property-wise subtyping for destructured bindings
Generally, property-wise subtyping is not safe because of aliasing.

```javascript
type X = { foo: string, bar: number };
type Y = { foo: string, bar?: number };
var x: X = { foo: "foo", bar: 0 };
var y: Y = x; // Unsound, even though Y < X
y.bar = undefined; // Here's why: (x : X) no longer holds
```

However, with destructured bindings, you don't end up with a reference
to the object on the RHS, so this kind of error isn't possible.

The reason-mangling is a bit unfortunate here. `mk_type_annotation` does
take a reason parameter, but discards it when there is a type
annotation and uses the reason generated from `convert` instead. For
objects, we end up with reason description of `"object type"`.

This commit adds a prefix to the generated reason, which is not too
intrusive, so we have `"destructured object type"` instead. This makes
it possible to identify the destructured binding in order to whitelist
the pattern for subtyping.