fix location of ClassDeclarations
Summary: we were inconsistent in the way we created locations for ClassDeclarations. in one place, we create a reason whose location is the identifier of the class; in others, the location is the entire class.

using the identifier's location is preferrable for error messages, because it doesn't point at your entire class, just its name. see the new test, for example. the entire class isn't shadowing the type `Foo`, only its identifier is.

there are surely other cases where we want to know the location of the entire class, but @bhosmer suggested that we should model that more like a class expression, where the name and the class definition are separate types. for now, we don't seem to rely on that, so improving error locations is the priority.

Reviewed By: @jeffmo

Differential Revision: D2359693