repositioning a reason should not unseal an object
Summary: we automatically seal objects when you try to access them from a different scope, where we currently loosly define "scope" as "same file". we do this by comparing the file from the object's reason with the file from the getter/setter's reason.

this falls apart if the reason is repositioned. what we actually want to do is keep track of where the object was defined, regardless of the reason.

this diff changes the `sealed` flag so that it carries that location around. I still use the location from the `GetT`/`SetT`/`GetElemT`/`SetElemT`/`ConstructorT`/whatever else, since we don't tend to reposition use types.

the `sealed` flag is also now extensible so that we can add something like `UnsealedInScope (Env_js.peek_frame ())`, which would enforce actual scopes (but breaks prototypical "classes", so I didn't do that now).

Reviewed By: @bhosmer

Differential Revision: D2368187