tweak position of things imported with require()
Summary: repositions the reason that comes out of a `require()` call to be the location of the `require()`, not the exports of the thing being required, when the thing being required uses CJS exports (`module.exports = ...` or `exports.Foo = ...` vs `export Foo`).

this brings parity between CJS exports and ES6 exports. if a file uses ES6 exports then the location in the requiring file is already the location of the `require()` (note how it creates an object using `CJSRequireT`'s reason in that case).

it also, IMO, improves the error messages significantly.

Reviewed By: @jeffmo

Differential Revision: D2368722