make --verbose greppable by pid
Summary: add pid to both lines of `--verbose` output so that grep will pick up both lines (could do `-A1` but this also makes it line up better)

also moved it out of the memoization check so that you can tell it's trying to recurse and bails out. this was super confusing to me.

Reviewed By: @gabelevi

Differential Revision: D2347787