make not operator invert booleans
Summary: the `!` operator previously created a generic `BoolT` and ignored the truthiness of the thing it was negating, so `!true` had unknown truthiness, which conflicts with things expecting a literal `false`.

this diff adds a `NotT` type which inverts the literal boolean value of the lower bound if one exists; if the lower bound is not a bool, then it creates an unknown-truthiness bool like it did before.

Reviewed By: @gabelevi

Differential Revision: D2384535