[hack] Implements `hh_server --daemon` on Windows.
Summary: In the absence of 'fork' on Windows, we added multiple entry points to `hh_server`.

All the platform-specific code is grouped in the module `Daemon`. This module export a function 'create_entry_point' to explicitely name alternate entry point. And a function `spawn` that allows to create an new instance of the current exectutable, that will execute a given alternate 'entry point'.

The function `spawn` replace the `Daemon.fork` function. Still, on Unix it is based on `Fork.fork` while on Windows, it is based on `Unix.create_process`.

This patch also introduces a function `Daemon.set_close_on_spawn`. It helps in managing file descriptors that should not leaks on `spawn`ed processes.
Closes https://github.com/facebook/hhvm/pull/6080

Reviewed By: @sgolemon

Differential Revision: D2367655