"fix" arrow function return type position
Summary: if an arrow function doesn't have an annotated return type, we infer one using a tvar, but the tvar is placed immediately //after// the `=>`, which is not where the annotation would've gone.

rather than try to position it in the "right" place, I made arrows with expression bodies (vs block bodies) use the entire expression's location, which is more useful. Note that the phantom return location for `foo => { return foo }` points at the space after `=>`, which is still wrong.

I can't get this to repro anywhere in practice since it's a tvar and we set the lower bound, so we always use the lower bound's location in errors, and we don't even show (most) OpenT's in traces, but I'm trying to break up a much bigger diff.

Reviewed By: @bhosmer

Differential Revision: D2387357