Warn when rest parameters are annotated with a tuple type
Before this change, annotating a rest parameter with a tuple type could
lead to an ugly exception, "Did not expect BoundT"

Instead, warn the user when this pattern occurs, because a tuple type
does not represent rest parameters well. The tuple type implies a fixed
number of elements, but there can be any number of rest parameters, by
definition.

While debugging this, I also noticed an issue with bound type parameters
in another rest parameter use case: `fn<T>(...xs: T)`. Currently, this
also throws the ugly exception, but it's a bit tricker to resolve
because the error is thrown from the unification that occurs before the
type param can be substituted.