Update waiting message to fit in 80 cols
Summary: On my terminal (which is 80 chars by default, like many), the "flow server still initializing" message looks like this:

{F23172443}

When the message is wider than the terminal, it can't update in place and it produces a new line each time the spinner spins. Shrink the message to fit in 80 cols.

Reviewed By: @jeffmo

Differential Revision: D2399132