make number of workers configurable
Summary: adds `server.max_workers` as a `.flowconfig` option and a `--max-workers` cli flag to `check`, `server` and `start`, to let you customize the number of workers to run. if not specified, it defaults to the number of processor cores (which was previously the hard-coded number of workers).

this is especially useful when using `--max-workers=0 --verbose`, since it prevents multiple processes' output from interleaving.

Reviewed By: @jeffmo

Differential Revision: D2403687