[flow parser] Add [optional/default-off] class-method decorator parsing support
Summary: Per discussions about having opt-in ignore/warn support for newer ES features that are still volatile or experimental, this adds support for the parser to take "parsing options" that can be read during parsing. More significantly, this uses this to add an `experimental_decorators` option -- which toggles support for parsing decorators on class methods.

Other than permitting/ignoring decorators, the effect on Flow from this diff is nothing right now since this option is off by default. In a small follow-up diff I'll plumb this new option through to flowconfig so that users can enable it in Flow themselves. We can then use this option-set to implement other early/experimental syntactic things to allow people to unblock their workflows when they wish to use a feature that Flow isn't ready to typecheck yet.

When deciding to implement parsing support for new features going forward, we'll have to make a decision about what the feature should do during inference. My current plan for decorators is to have users specify either `ignore` or `warn` as the value of the flowconfig option. This will let users choose if they wish for Flow to just get out of the way when encountering a decorator (useful for innocuous things that the user knows won't affect typechecking like loggers or memoizers), or if they want to be warned about the fact that decorators can interfere with the validity of Flow's understanding of the code (and perhaps require the user to place an explicit `/*FlowIgnore*/` in the code).

Reviewed By: @nmote

Differential Revision: D2407416