Respect flowconfig [ignore] always when searching for node modules
Summary: Before this change, we ignored the path for a `package.json` file based
on the [ignore] section of the flowconfig, but we still tried to find an
`index.js` file regardless. Later on, Flow would complain that the file
could not be found.

Before this change, for the provided test case, users would get an
error:

```
test.js:3:17,21: index.js
Required module not found

Found 1 error
```

After this change, we succeed in using the libdef, so we might see an
expected error:

```
test.js:5:2,8: number
This type is incompatible with
test.js:5:12,17: string

Found 1 error
```

cc @jeffmo
Closes https://github.com/facebook/flow/pull/778

Reviewed By: @gabelevi

Differential Revision: D2414241

Pulled By: @mroch