Widen type for shadow-bound var in merged-env scenarios
Before this change, the included test case did not produce any errors,
when in fact the returned type may be widened.

In order to widen the type of x, we need to add an entry to the
changeset, so the specific value for that binding in the merged-in env
is added as a lower bound in the merged-to env.

For "normal" assignment, e.g., `x = ""`, we go through `set_var` which
does add a changeset entry. A redeclaration, however, bypasses `set_var`.