Implement class method generators
Class methods go through a quite different, and a bit more convoluted
process. The hack I was using for function declarations where I swapped
out the inferred return type with a Generator typeapp no longer applies.

This is a good thing, because it forced me to replace the hack with
something better. Like async functions, the fancy generator return type
is established in the Return statement handler directly. If there is no
explicit return, we default to a Generator typeapp where R=void.

I intentionally left the capacity for async=true/generator=true var
scopes. Babel, via regenerator[1], seems to understand `async function*`
using the AsyncIterator[2] proposal.

For now, the parser will reject async generator functions, but we might
choose to support async iterators in the future, so having the capacity
to do so in VarScope seems right.

1. https://github.com/facebook/regenerator/blob/9423b4ade99173b603430a151599cf8637166f16/test/async.es6.js#L259
2. https://github.com/zenparsing/async-iteration/#async-generator-functions